+++ src/containers/find-idpw/FindIdPw.tsx
@@ -1,13 +1,13 @@
-import FindId from './FindId'
-// import FindPwd from './FindPwd'
+import IdForm from './IdForm'
+// import PwForm from './PwForm'
 
 export default function FindIdPwd() {
   return (
     <div className="border px-5 py-[26px] border-b-0 border-solid border-[#ddd] bg-[#fff] px-[13px]">
       {/* 아이디 찾기 */}
-      <FindId />
+      <IdForm />
       {/* 비밀번호 찾기 */}
-      {/* <FindPwd /> */}
+      {/* <PwForm /> */}
     </div>
   )
 }
